Making code pass pytype in Python 3.10, 3.11 and 3.12 ; some cleanup.
